requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.14
django-htmx==1.12.2
djangorestframework==3.14.0
idna==3.4
Pillow==9.2.0
Pint==0.19.2
psycopg2-binary==2.9.3
pyTelegramBotAPI==4.7.0
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
whitenoise==6.2.0
